Major refactor ✨
Though this was done more than a year ago (in October 2023), so I don't fully remember the changes I had made, the main goal was to make the project structure cleaner, while also allowing this to be used both as a desktop as well as a mobile app. For this purpose, I changed the project structure, moving the logic (downloading data from storj, searching a word in a dictionary file) to the pkg folder, and the desktop and website app packages to /cmd.

To compile the website (I'm pretty sure the desktop app doesn't work right now), use go build cmd/website/main.go