Actually, remove wails app ._.
Silly of me - but the previous "major refactor" commit was in reality fully made 1.5 year ago. Back then I didn't know lower level programming languages or other GUI frameworks, so I thought GO Wails was the best option. However, it doesn't work on mobile, it isn't as mature as other ui frameworks, and it doesn't look and behave like a native (in linux a GTK or QT) app. For this reason, I think the future of IxaLang is a QT app using an XMLListModel.

In fact, there is another 2 major issues with this app/site as it stands: The website won't work well in production (with many users) due the file bottleneck, and (most importantly!) the dictionaries are currently not included in GitHub, nor can be downloaded. Hence, the plan is to rewrite the app in C/C++ (using web assembly on the browser to download dictionary files).