Nearly made WikDict work with a WASM SQLite worker
as you can see, i wrote quite a lot of code, lol. i didn't expect it'd be so difficult. but tbh i got greedy and thought i might get freedict to work and create a native app in the future. so i tried to keep the JS implementation as runtime and UI independent as possible.

also write some terrible PHP cause python is stupid and i can't use headers. this still doesn't fix the OPFS issue but since we are using a different implementation (no. 2 in SQLite docs // ashimoto's VFS) it still seems to work. up to the statement that's invalid that is.